Split code from cocoa_webkit.hh
